Offset type Long

+++ app/src/main/java/one/mixin/android/api/service/AssetService.kt
@@ -28,7 +28,7 @@ interface AssetService {
 
     @GET("assets/{id}/snapshots")
     fun snapshots(@Path("id") id: String,
-        @Query("offset") offset: Int = 0,
+        @Query("offset") offset: Long = 0L,
         @Query("limit") limit: Int = 100
     ): Call<MixinResponse<List<Snapshot>>>
 
@@ -48,7 +48,7 @@ interface AssetService {
     fun addresses(@Path("id") id: String): Call<MixinResponse<List<Address>>>
 
     @GET("snapshots")
-    fun allSnapshots(@Query("offset") offset: Int = 0,
+    fun allSnapshots(@Query("offset") offset: Long = 0L,
         @Query("limit") limit: Int = 100
     ): Call<MixinResponse<List<Snapshot>>>
